+++ src/CarouselItem.js
@@ -1,5 +1,6 @@
 import classNames from 'classnames';
 import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 
 import TransitionEvents from './utils/TransitionEvents';
@@ -8,12 +9,12 @@ import TransitionEvents from './utils/TransitionEvents';
 // not wait until transition end to trigger continuing animations.
 
 const propTypes = {
-  direction: React.PropTypes.oneOf(['prev', 'next']),
-  onAnimateOutEnd: React.PropTypes.func,
-  active: React.PropTypes.bool,
-  animateIn: React.PropTypes.bool,
-  animateOut: React.PropTypes.bool,
-  index: React.PropTypes.number,
+  direction: PropTypes.oneOf(['prev', 'next']),
+  onAnimateOutEnd: PropTypes.func,
+  active: PropTypes.bool,
+  animateIn: PropTypes.bool,
+  animateOut: PropTypes.bool,
+  index: PropTypes.number,
 };
 
 const defaultProps = {