+++ src/Glyphicon.js
@@ -1,14 +1,15 @@
 import classNames from 'classnames';
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import { bsClass, getClassSet, prefix, splitBsProps }
   from './utils/bootstrapUtils';
 
 const propTypes = {
   /**
-   * An icon name. See e.g. http://getbootstrap.com/components/#glyphicons
+   * An icon name without "glyphicon-" prefix. See e.g. http://getbootstrap.com/components/#glyphicons
    */
-  glyph: React.PropTypes.string.isRequired,
+  glyph: PropTypes.string.isRequired,
 };
 
 class Glyphicon extends React.Component {