+++ src/Modal.js
@@ -4,6 +4,7 @@ import ownerDocument from 'dom-helpers/ownerDocument';
 import canUseDOM from 'dom-helpers/util/inDOM';
 import getScrollbarSize from 'dom-helpers/util/scrollbarSize';
 import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import BaseModal from 'react-overlays/lib/Modal';
 import isOverflowing from 'react-overlays/lib/utils/isOverflowing';
@@ -28,17 +29,17 @@ const propTypes = {
    * Include a backdrop component. Specify 'static' for a backdrop that doesn't
    * trigger an "onHide" when clicked.
    */
-  backdrop: React.PropTypes.oneOf(['static', true, false]),
+  backdrop: PropTypes.oneOf(['static', true, false]),
 
   /**
    * Close the modal when escape key is pressed
    */
-  keyboard: React.PropTypes.bool,
+  keyboard: PropTypes.bool,
 
   /**
    * Open and close the Modal with a slide and fade animation.
    */
-  animation: React.PropTypes.bool,
+  animation: PropTypes.bool,
 
   /**
    * A Component type that provides the modal content Markup. This is a useful
@@ -53,55 +54,61 @@ const propTypes = {
    * Generally this should never be set to false as it makes the Modal less
    * accessible to assistive technologies, like screen-readers.
    */
-  autoFocus: React.PropTypes.bool,
+  autoFocus: PropTypes.bool,
 
   /**
    * When `true` The modal will prevent focus from leaving the Modal while
    * open. Consider leaving the default value here, as it is necessary to make
    * the Modal work well with assistive technologies, such as screen readers.
    */
-  enforceFocus: React.PropTypes.bool,
+  enforceFocus: PropTypes.bool,
+
+  /**
+   * When `true` The modal will restore focus to previously focused element once
+   * modal is hidden
+   */
+  restoreFocus: PropTypes.bool,
 
   /**
    * When `true` The modal will show itself.
    */
-  show: React.PropTypes.bool,
+  show: PropTypes.bool,
 
   /**
    * A callback fired when the header closeButton or non-static backdrop is
    * clicked. Required if either are specified.
    */
-  onHide: React.PropTypes.func,
+  onHide: PropTypes.func,
 
   /**
    * Callback fired before the Modal transitions in
    */
-  onEnter: React.PropTypes.func,
+  onEnter: PropTypes.func,
 
   /**
    * Callback fired as the Modal begins to transition in
    */
-  onEntering: React.PropTypes.func,
+  onEntering: PropTypes.func,
 
   /**
    * Callback fired after the Modal finishes transitioning in
    */
-  onEntered: React.PropTypes.func,
+  onEntered: PropTypes.func,
 
   /**
    * Callback fired right before the Modal transitions out
    */
-  onExit: React.PropTypes.func,
+  onExit: PropTypes.func,
 
   /**
    * Callback fired as the Modal begins to transition out
    */
-  onExiting: React.PropTypes.func,
+  onExiting: PropTypes.func,
 
   /**
    * Callback fired after the Modal finishes transitioning out
    */
-  onExited: React.PropTypes.func,
+  onExited: PropTypes.func,
 
   /**
    * @private
@@ -116,8 +123,8 @@ const defaultProps = {
 };
 
 const childContextTypes = {
-  $bs_modal: React.PropTypes.shape({
-    onHide: React.PropTypes.func,
+  $bs_modal: PropTypes.shape({
+    onHide: PropTypes.func,
   }),
 };