+++ src/ModalHeader.js
@@ -1,5 +1,6 @@
 import classNames from 'classnames';
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import { bsClass, getClassSet, splitBsProps } from './utils/bootstrapUtils';
 import createChainedFunction from './utils/createChainedFunction';
@@ -12,19 +13,19 @@ const propTypes = {
    * button. It is used for Assistive Technology when the label text is not
    * readable.
    */
-  'aria-label': React.PropTypes.string,
+  'aria-label': PropTypes.string,
 
   /**
    * Specify whether the Component should contain a close button
    */
-  closeButton: React.PropTypes.bool,
+  closeButton: PropTypes.bool,
 
   /**
    * A Callback fired when the close button is clicked. If used directly inside
    * a Modal component, the onHide will automatically be propagated up to the
    * parent Modal `onHide`.
    */
-  onHide: React.PropTypes.func,
+  onHide: PropTypes.func,
 };
 
 const defaultProps = {
@@ -33,8 +34,8 @@ const defaultProps = {
 };
 
 const contextTypes = {
-  $bs_modal: React.PropTypes.shape({
-    onHide: React.PropTypes.func,
+  $bs_modal: PropTypes.shape({
+    onHide: PropTypes.func,
   }),
 };
 
@@ -65,7 +66,7 @@ class ModalHeader extends React.Component {
             type="button"
             className="close"
             aria-label={label}
-            onClick={createChainedFunction(modal.onHide, onHide)}
+            onClick={createChainedFunction(modal && modal.onHide, onHide)}
           >
             <span aria-hidden="true">
               &times;