+++ src/NavDropdown.js
@@ -1,5 +1,6 @@
 import classNames from 'classnames';
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import Dropdown from './Dropdown';
 import splitComponentProps from './utils/splitComponentProps';
@@ -9,15 +10,15 @@ const propTypes = {
   ...Dropdown.propTypes,
 
   // Toggle props.
-  title: React.PropTypes.node.isRequired,
-  noCaret: React.PropTypes.bool,
-  active: React.PropTypes.bool,
+  title: PropTypes.node.isRequired,
+  noCaret: PropTypes.bool,
+  active: PropTypes.bool,
 
   // Override generated docs from <Dropdown>.
   /**
    * @private
    */
-  children: React.PropTypes.node,
+  children: PropTypes.node,
 };
 
 class NavDropdown extends React.Component {