+++ src/ResponsiveEmbed.js
@@ -1,5 +1,6 @@
 import classNames from 'classnames';
-import React, { cloneElement, PropTypes } from 'react';
+import React, { cloneElement } from 'react';
+import PropTypes from 'prop-types';
 import warning from 'warning';
 
 import { bsClass, getClassSet, prefix, splitBsProps }