+++ src/Tab.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import TabContainer from './TabContainer';
 import TabContent from './TabContent';
@@ -7,14 +8,14 @@ import TabPane from './TabPane';
 const propTypes = {
   ...TabPane.propTypes,
 
-  disabled: React.PropTypes.bool,
+  disabled: PropTypes.bool,
 
-  title: React.PropTypes.node,
+  title: PropTypes.node,
 
   /**
    * tabClassName is used as className for the associated NavItem
    */
-  tabClassName: React.PropTypes.string
+  tabClassName: PropTypes.string
 };
 
 class Tab extends React.Component {