+++ src/Table.js
@@ -1,15 +1,16 @@
 import classNames from 'classnames';
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import { bsClass, getClassSet, prefix, splitBsProps }
   from './utils/bootstrapUtils';
 
 const propTypes = {
-  striped: React.PropTypes.bool,
-  bordered: React.PropTypes.bool,
-  condensed: React.PropTypes.bool,
-  hover: React.PropTypes.bool,
-  responsive: React.PropTypes.bool,
+  striped: PropTypes.bool,
+  bordered: PropTypes.bool,
+  condensed: PropTypes.bool,
+  hover: PropTypes.bool,
+  responsive: PropTypes.bool,
 };
 
 const defaultProps = {